Split RPC tool tests by domain
